clocky: clock-atom now array-map with keyword keys
The order of the clock-atom is important because the
clock needs to be rendered with the largest arcs
first, so hours, then minutes, etc. So we should use
an array-map to hold the contents of this data instead
of just a map. Note that this makes some light assumptions
on the inner workings of unify.

Also replaced the string keys with keywords for slightly
cleaner looking code, at the expense of calling (name)
a couple of times.