clocky: unify replaced with for, clock-atom now map
Since order of processing is important, replaced unify
with a for comprehension over the 4 keys. This means
we can also drop the array-map and use a simple map.

Note however that this version no longer works because
the bind! macro no longer seems to be watching the
clock-atom when it is not part of a unify. I'll
followup with a separate commit that fixes this.